Add ui_common.cpp
